requirements.txt
twitchio==2.9.0
aiohttp==3.8.0
requests==2.31.0
streamlit==1.28.0
plotly==5.17.0
pandas==2.0.0
numpy==1.24.0
python-dotenv==1.0.0

# For BERT
torch==2.1.0
transformers==4.35.0
scikit-learn==1.3.0